Refactor color management into modular ColorManager module
